backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.0
openai==1.54.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
playwright==1.49.0
httpx==0.24.0
